Fix signed integer overflow (#572)
In Issue #571, I encountered a bus error.
Using AddressSanitizer, I traced the error to a negative memory index
caused by a 32 bit signed integer overflowing the bounds and requesting
memory from a negative index.

Swap out `int` for `long` so that more total memory can be addressed and
allow larger problem sizes.

+++ smt/src/rmts/rmtc.hpp
@@ -18,7 +18,7 @@ private:
   void find_interval(int ix, int num, double x, int * index, double * xbar);
   int * nelem_list;
   int * nterm_list;
-  int nelem, nterm;
+  long nelem, nterm;
 };
 
 #endif
